added class structure for the mini apps
Okay, so this is a bit of a lot, but basically there are chapters which
are interactive (mini apps) instead of just a playing movie. So I
created a class called miniHandler, this one should handle which apps
are allowed to play and closes and opens them etc. mainMini is a class
that containts the stuff all the apps would possibly need. All mini
apps derive their goods from the mainMini.

Still need to figure out how to add these to the chapter list, now only
contains movies. the apps should say after wich chapter they fit in.

In the end the play all should include both chapters and mini apps,
mini apps have a way of saying it's the end of their time and the show
should move on.