added some simple controls so switching miniApps/video's will go a bit nicer
stops an app when active if you're selecting a movie and vice versa.

Still have the need for some object that contains both miniApps and
video's in their respective order and tells me basic things like what
type they are, which one is active and should function as a
intermediator for the controlls.