attempt at XML file 300 char chop for sending over MPE network for checking
want to send the XML file over the network as a string, so each app can
confirm this is the same as what they've got in their directory. Hoever
MPE doesn't seem to like strings bigger than 500 or so characters.
Chopping the XML string up seems to give some strange results.. will
work on it tomorrow.