Bug fixes on toggle buttons and the playAll function
Think I fixed the play all function, before it was first come first
serve when it came to triggering the next button when the play all was
on. Now on the incoming side it has a counter that diables the next
action for 100 millisec when it gets one. So you can't have two apps
next'ing right after each other.

Also fixed some weird stuff that happens at the syphon and playall
toggle buttons. For some reason their counter (because the GUI sends a
double value I somehow can't filter.. ) doesn't get updated so you
needed to press a button with a different labelname before pressing the
other or it would be ignored. Now that's sorta fixed.