plaats_origine looping works better
now it calculates the shortest distance between the current frame and
the frame it has to go to, even if this means going through the
end/start point.

still fiddeling around with the codecs though.. the artwork seems to
work fine (because of the small size I think) but the background is
still shakey.