basically finisihed interview_bezoekers chapter
Not yet included variable amounts of videos being displayed, but yeah
it has the XML in out point function, video's are being handled
properly put in a viewport and rotated.
- Every app loads all films and plays the current one, could do with a
check if the video is in the apps viewport.

For the rest, added scrolling for the chapter GUI, fixed some bugs,
added an exit function for all miniApps, did some cleaning in
miniHandler, replacing the strings with string variables might make
things a bit less buggy.